优化认证

+++ app.py
@@ -114,22 +114,28 @@ def send_file_detail():
 
 @app.route('/toggleBookmark')
 def toggle_bookmark():
-    path = request.args.get("path")
-    bookmark_flag_file = resource_path('') + 'preview/' + path.replace("/", "_") + '.bookmark'
-    state = os.path.exists(bookmark_flag_file)
-    if state:
-        os.remove(bookmark_flag_file)
+    if check_client_ip(request.remote_addr):
+        path = request.args.get("path")
+        bookmark_flag_file = resource_path('') + 'preview/' + path.replace("/", "_") + '.bookmark'
+        state = os.path.exists(bookmark_flag_file)
+        if state:
+            os.remove(bookmark_flag_file)
+        else:
+            with open(bookmark_flag_file, 'w') as fp:
+                fp.write("This is a Bookmark file!")
+        return "成功取消标记" if state else "成功标记为看过"
     else:
-        with open(bookmark_flag_file, 'w') as fp:
-            fp.write("This is a Bookmark file!")
-    return "成功取消标记" if state else "成功标记为看过"
+        return "Permission Denied"
 
 
 @app.route("/getFile/<file_name>")
 def get_file(file_name):
-    # url中加一个文件名避免播放器不知道视频文件名
-    path = request.args.get("path")
-    return send_file(root + path, as_attachment=True, attachment_filename=path[path.rindex("/") + 1:], conditional=True)
+    if check_client_ip(request.remote_addr):
+        # url中加一个文件名避免播放器不知道视频文件名
+        path = request.args.get("path")
+        return send_file(root + path, as_attachment=True, attachment_filename=path[path.rindex("/") + 1:], conditional=True)
+    else:
+        return "Permission Denied", 404
 
 
 @app.route("/getVideoPreview")
@@ -176,13 +182,16 @@ def get_known_mime(mime_type=''):
 
 @app.route("/remoteDownload")
 def add_remote_download():
-    url = request.args.get("url").replace("__and__", "&")
-    name = request.args.get("name")
-    l = name.rindex("/")
-    import subprocess
-    cmd = 'wget -P %s -O %s "%s"' % (root + name[:l], name[l + 1:], url)
-    subprocess.call(cmd, shell=True)
-    return "成功添加离线任务:" + name
+    if check_client_ip(request.remote_addr):
+        url = request.args.get("url").replace("__and__", "&")
+        name = request.args.get("name")
+        l = name.rindex("/")
+        import subprocess
+        cmd = 'wget -P %s -O %s "%s"' % (root + name[:l], name[l + 1:], url)
+        subprocess.call(cmd, shell=True)
+        return "成功添加离线任务:" + name
+    else:
+        return "Permission Denied"
 
 
 def check_client_ip(ip):
@@ -197,7 +206,7 @@ def user_login():
     with open(resource_path('')+"user.json", 'r') as f:
         j = json.loads(f.read())
         try:
-            if j[name] == psw:
+            if j[name] == psw and not check_client_ip(request.remote_addr):
                 j['ip'].append(request.remote_addr)
                 with open(resource_path('') + "user.json", 'w') as f1:
                     f1.write(json.dumps(j))
